// app/app/api/generations/schema.ts
import z from "zod";
import { PaginationSchema } from "../pagination";

const GenerationsStatus = z.enum(["PENDING", "SUCCESS", "ERROR"])

export enum Animals {
  Wolf = "wolf",
  Cat = "cat",
  Chicken = "chicken",
  Deer = "deer",
  Hedgehog = "hedgehog",
  Owl = "owl",
  Warthog = "warthog"
}

export const AnimalSchema = z.enum(Animals)

export const GenerationSchema = z.object({
  jobId: z.string(),
  groupId: z.string(),
  status: GenerationsStatus,
  result: z.string().optional(),
  error: z.string().optional()
})

export type Generation = z.infer<typeof GenerationSchema>

export const GenerationsIndexResponseSchema = PaginationSchema(GenerationSchema);
export type GenerationsIndexResponse = z.infer<typeof GenerationsIndexResponseSchema>

export const GenerationsCreateInputSchema = z.object({
  numImages: z.number().min(0).max(10),
  animal: z.string(),
})

export type GenerationsCreateInput = z.infer<typeof GenerationsCreateInputSchema>

export const GenerationsCreateResponseSchema = z.array(GenerationSchema)

export type GenerationsCreateResponse = z.infer<typeof GenerationsCreateResponseSchema>

export const GenerationsIndexQuerySchema = z.object({
  offset: z.coerce.number().min(0),
  pageSize: z.coerce.number().min(0).max(25)
})

export type GenerationsIndexQuery = z.infer<typeof GenerationsIndexQuerySchema>